% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.R
\name{dv_dataset}
\alias{dv_dataset}
\title{Get dataset details by DOI}
\usage{
dv_dataset(doi)
}
\arguments{
\item{doi}{Dataset persistent ID (e.g. "doi:10.7910/DVN/EXAMPLE")}
}
\value{
tibble: name, doi, description, publisher, published_at,
file_name, file_id, file_size, file_type (one row per file)
}
\description{
Get dataset details by DOI
}
